重构现有代码，添加XmlAttributeThirdPartyCallback相关文件。
