+++ QYWXLocalDebug.csproj
@@ -87,7 +87,20 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="WXSDK\WXBizMsgCrypt.cs" />
-    <Compile Include="XMLRequest\XMLText.cs" />
+    <Compile Include="XmlAttributeThirdPartyCallback.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="XmlAttributeThirdPartyCallback.Designer.cs">
+      <DependentUpon>XmlAttributeThirdPartyCallback.cs</DependentUpon>
+    </Compile>
+    <Compile Include="XMLRequest\BoundXmlBase.cs" />
+    <Compile Include="XMLRequest\BoundXmlThirdPartyCallback.cs" />
+    <Compile Include="XMLRequest\BoundXmlText.cs" />
+    <Compile Include="XMLRequest\RequestXmlBase.cs" />
+    <Compile Include="XMLRequest\RequestXmlEventKey.cs" />
+    <Compile Include="XMLRequest\RequestXmlThirdPartyCallback.cs" />
+    <Compile Include="XMLRequest\RequestXmlText.cs" />
+    <Compile Include="XMLRequest\BoundXmlEventKey.cs" />
     <Compile Include="XmlAttributeBase.cs">
       <SubType>UserControl</SubType>
     </Compile>
@@ -120,6 +133,9 @@
     </EmbeddedResource>
     <EmbeddedResource Include="XmlAttributeText.resx">
       <DependentUpon>XmlAttributeText.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="XmlAttributeThirdPartyCallback.resx">
+      <DependentUpon>XmlAttributeThirdPartyCallback.cs</DependentUpon>
     </EmbeddedResource>
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
