+++ XmlAttributeBase.cs
@@ -28,7 +28,7 @@ namespace QYWXLocalDebug
         }
 
         //XMLText自定义的XMLChanged事件，用于刷新发送消息的xml格式，实时刷新
-       protected  void XML_AttributeChanged(RequestXMLBase xml)
+       protected  void XML_AttributeChanged(RequestXmlBase xml)
         {
             if (XMLTextBox != null)
                 XMLTextBox.Text = xml.GetXML();
