+++ XmlAttributeSubscribe.cs
@@ -32,14 +32,14 @@ namespace QYWXLocalDebug
 
             txt_MsgType.Text = "event";
             //模拟微信发送消息
-            BoundXMLEvent_Key XML = new BoundXMLEvent_Key();
-            XML.Add(txt_ToUserName, BoundXMLEvent_Key.ToUserName);
-            XML.Add(txt_FromUserName, BoundXMLEvent_Key.FromUserName);
-            XML.Add(txt_CreateTime, BoundXMLEvent_Key.CreateTime);
-            XML.Add(txt_MsgType, BoundXMLEvent_Key.MsgType);
-            XML.Add(txt_Event, BoundXMLEvent_Key.Event);
+            BoundXmlEventKey XML = new BoundXmlEventKey();
+            XML.Add(txt_ToUserName, BoundXmlEventKey.ToUserName);
+            XML.Add(txt_FromUserName, BoundXmlEventKey.FromUserName);
+            XML.Add(txt_CreateTime, BoundXmlEventKey.CreateTime);
+            XML.Add(txt_MsgType, BoundXmlEventKey.MsgType);
+            XML.Add(txt_Event, BoundXmlEventKey.Event);
             //XML.Add(txt_EventKey, BoundXMLEvent_Key.EventKey);
-            XML.Add(txt_AgentID, BoundXMLEvent_Key.AgentID);
+            XML.Add(txt_AgentID, BoundXmlEventKey.AgentID);
 
             XML.IniNotice();
